fix: resolve Log4j2 stack trace and DuckDB type mapping issues
- Fixed ClassCastException by changing parsed_event_logs.status to INTEGER and matching EventLogStatus enum
- Enhanced log4j2.xml with %xEx for extended stack traces
- Updated JacksonEventParser logging to be more explicit about parsing errors
- Introduced 'build-frontend' Maven profile to optimize backend build times

+++ src/main/java/com/spark/insight/model/EventLogStatus.java
@@ -5,16 +5,16 @@ import lombok.Getter;
 
 @Getter
 public enum EventLogStatus {
-    PROCESSING((byte) 0L, "Parsing"),
-    IMPORTING((byte) 1, "Importing"),
-    SUCCESS((byte) 2, "Success"),
-    FAILED((byte) 3, "Failed");
+    PROCESSING(0, "Parsing"),
+    IMPORTING(1, "Importing"),
+    SUCCESS(2, "Success"),
+    FAILED(3, "Failed");
 
     @EnumValue
-    private final Byte code;
+    private final Integer code;
     private final String description;
 
-    EventLogStatus(Byte code, String description) {
+    EventLogStatus(Integer code, String description) {
         this.code = code;
         this.description = description;
     }
